requirements.txt
pandas
python-dotenv
PyYAML
requests
selenium
webdriver-manager
psutil